refactor: split batch.rs into batch/ directory + dedup read logic (CQ-8/CQ-9)
2844-line monolith → 4 focused files:
- batch/mod.rs (337): BatchContext, cmd_batch main loop
- batch/commands.rs (643): BatchInput/BatchCmd parsing + dispatch router
- batch/handlers.rs (1173): 23 dispatch_* handler functions
- batch/pipeline.rs (560): pipe chaining, fan-out, pipeline execution

Extract shared read functions into commands/read.rs:
- validate_and_read_file: path traversal + size check
- build_file_note_header: note injection for file reads
- build_focused_output: focused read rendering + type deps

Both CLI and batch read paths now call shared core, eliminating
~200 lines of duplicated logic.
