+++ CONTRIBUTING.md
@@ -89,7 +89,11 @@ src/
     mod.rs      - Argument parsing, command dispatch
     commands/   - Command implementations
       mod.rs, query.rs, index.rs, stats.rs, graph.rs, init.rs, doctor.rs, notes.rs, reference.rs, similar.rs, explain.rs, diff.rs, drift.rs, trace.rs, impact.rs, impact_diff.rs, test_map.rs, context.rs, resolve.rs, dead.rs, gc.rs, gather.rs, project.rs, audit_mode.rs, read.rs, stale.rs, related.rs, where_cmd.rs, scout.rs, onboard.rs, convert.rs, review.rs, ci.rs, health.rs, suggest.rs, deps.rs
-    batch.rs    - Batch mode: persistent Store + Embedder, stdin commands, JSONL output, pipeline syntax
+    batch/      - Batch mode: persistent Store + Embedder, stdin commands, JSONL output, pipeline syntax
+      mod.rs      - BatchContext, vector index builder, main loop
+      commands.rs - BatchInput/BatchCmd parsing, dispatch router
+      handlers.rs - Handler functions (one per command)
+      pipeline.rs - Pipeline execution (pipe chaining via `|`)
     config.rs   - Configuration file loading
     display.rs  - Output formatting, result display
     files.rs    - File enumeration, lock files, path utilities
