+++ PROJECT_CONTINUITY.md
@@ -2,19 +2,22 @@
 
 ## Right Now
 
-**v0.12.12 audit complete (P1-P3).** 2026-02-21.
+**batch.rs split + P4 audit.** 2026-02-21.
 
-- P1: 12 fixes merged (#459) — security, correctness, data safety, error handling
-- P2: 18 fixes merged (#460) — BatchContext caching, N+1 queries, API design, robustness. Also renamed `gpu-search` → `gpu-index`
-- P3: 31 fixes merged (#461) — docs (9), error handling (5), observability (4), API design (5), robustness (5), algorithm/platform/perf (5), test coverage (2), doc comments (3). 8 non-issues triaged.
-- P4: 18 findings deferred (tests, extensibility, design)
-- Binary updated, index rebuilt (3984 chunks, 4276 type edges)
+- P1-P3 merged (#459, #460, #461)
+- P4: 18 fixes on branch `fix/p4-audit-findings`
+- batch.rs split: 2844-line monolith → batch/ directory (4 files) + CQ-8/CQ-9 read dedup
+- Ready to commit split work, then PR
 
 Audit triage: `docs/audit-triage.md`
 
 ## Pending Changes
 
-None.
+Uncommitted on `fix/p4-audit-findings`:
+- `src/cli/batch.rs` deleted → `src/cli/batch/{mod,commands,handlers,pipeline}.rs`
+- `src/cli/commands/read.rs` rewritten with shared core functions
+- `src/health.rs` cosmetic comment fix
+- `CONTRIBUTING.md` architecture section updated
 
 ## Parked
 
@@ -25,7 +28,7 @@ None.
 - **Phase 8**: Security (index encryption)
 - **ref install** — deferred, tracked in #255
 - **Query-intent routing** — auto-boost ref weight when query mentions product names
-- **P4 audit findings** — 18 findings deferred (issues TBD)
+- **P4 audit findings** — 18 findings fixed on `fix/p4-audit-findings`
 - **resolve_target test bias** — ambiguous names resolve to test functions over production code. Not blocking, but `cqs related foo` may pick `test_foo_bar` instead of `foo`. Fix: prefer non-test chunks in resolve_target.
 
 ## Open Issues
@@ -49,7 +52,7 @@ None.
 - HNSW index: chunks only (notes use brute-force SQLite search)
 - Multi-index: separate Store+HNSW per reference, parallel rayon search, blake3 dedup
 - 9 languages (Rust, Python, TypeScript, JavaScript, Go, C, Java, SQL, Markdown)
-- Tests: 1020 pass + 28 ignored
+- Tests: 1037 pass + 31 ignored
 - CLI-only (MCP server removed in PR #352)
 - Source layout: parser/, hnsw/, impact/ are directories (impact split in PR #402)
 - convert/ module (7 files) behind `convert` feature flag
